kvcoord: move condensableSpanSet to its own file
... in anticipation of it being used by a second interceptor.

Release note: None
